Added example prefab class
It can be used as a model to base the rest of the "scripts" off of.
I also changed the file importers to take const char*s (rather than
normal char*s) for their filepath parameters. We didn't want to change
those values in the first place, and putting it this way allows them to
be used with constant variables (like those returned by a std::string's
c_str() function, for example).