Reworked keyboard input
Using one map rather than individual bools for each key makes the code
much more scalable and easy to use in a variety of situations. I also
added the ability for the camera to move up and down (with space and
left shift) in addition to WASD planar movement. I'm not sure how often
that'll be used once we tie the camera to the player, but it's nice to
have while we control a disembodied spectator cam.

Current points of interest: Using an unordered_map would probably be
preferable, but the program crashes at runtime when its member is
changed from a map to an unordered_map. In its current form, the program
suffers heap corruption when exiting. Neither of these issues are
particularly problematic, but both of them are worth notice.